monty/requirements.txt
bottle
pydantic
psycopg2
numpy
matplotlib
waitress
pyjwt
requests
python-dateutil